requirements.txt
flask
psycopg2-binary
python-dotenv
werkzeug
os
time
datetime